requirements.txt
aiohttp>=3.9
feedparser>=6.0
gspread>=6.0
gspread_asyncio>=2.0
google-auth>=2.0
python-dateutil>=2.9
python-dotenv>=1.0
pytz>=2024.1
